Add ScheduleServiceImpl(not Test)

+++ src/main/java/com/dreamsecurity/shopface/service/ScheduleServiceImpl.java
@@ -2,10 +2,15 @@ package com.dreamsecurity.shopface.service;
 
 import com.dreamsecurity.shopface.domain.Branch;
 import com.dreamsecurity.shopface.domain.Schedule;
+import com.dreamsecurity.shopface.dto.employ.EmployListResponseDto;
 import com.dreamsecurity.shopface.dto.schedule.ScheduleAddRequestDto;
 import com.dreamsecurity.shopface.dto.schedule.ScheduleEditRequestDto;
 import com.dreamsecurity.shopface.dto.schedule.ScheduleListResponseDto;
 import com.dreamsecurity.shopface.dto.schedule.ScheduleResponseDto;
+import com.dreamsecurity.shopface.enums.ScheduleColor;
+import com.dreamsecurity.shopface.enums.ScheduleState;
+import com.dreamsecurity.shopface.repository.EmployRepository;
+import com.dreamsecurity.shopface.repository.OccupationRepository;
 import com.dreamsecurity.shopface.repository.ScheduleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,19 +23,64 @@ import java.util.stream.Collectors;
 @Service
 public class ScheduleServiceImpl implements  ScheduleService {
     private final ScheduleRepository scheduleRepository;
+    private final EmployRepository employRepository;
+    private final OccupationRepository occupationRepository;
 
     @Transactional
     @Override
     public Long addSchedule(ScheduleAddRequestDto requestDto) {
-        return scheduleRepository.save(requestDto.toEntity()).getNo();
+        if (requestDto.getWorkStartTime().isBefore(requestDto.getWorkEndTime())) {
+            List<EmployListResponseDto> entity = employRepository.findByMemberIdAndBranchNo(requestDto.getMember().getId(), requestDto.getBranch().getNo());
+            if (!entity.isEmpty()) {
+                // 스케줄 중복 검사
+                List<ScheduleListResponseDto> oldSchedules = scheduleRepository.findAllByTodayAndMemberIdAndBranchNo(requestDto.getWorkStartTime(),
+                        requestDto.getMember().getId(),
+                        requestDto.getBranch().getNo());
+                for (ScheduleListResponseDto old : oldSchedules) {
+                    if ((requestDto.getWorkStartTime() == old.getWorkStartTime())
+                            || (requestDto.getWorkEndTime() == old.getWorkEndTime())) {
+                        new IllegalArgumentException("동시간대에 다른 스케줄이 존재합니다.");
+                    } else if (requestDto.getWorkEndTime().isBefore(old.getWorkStartTime())) {
+                        if (requestDto.getWorkStartTime().isBefore(old.getWorkEndTime())) {
+                            continue;
+                        } else {
+                            new IllegalArgumentException("다른스케줄이 존재하여 등록할 수 없습니다.");
+                        }
+                    } else if (requestDto.getWorkEndTime().isAfter(old.getWorkStartTime())) {
+                        if (requestDto.getWorkStartTime().isAfter(old.getWorkEndTime())) {
+                            continue;
+                        } else {
+                            new IllegalArgumentException("등록하려는 시간대에 다른스케줄이 존재하여 등록할 수 없습니다.");
+                        }
+                    } else {
+                        new IllegalArgumentException("등록하려는 시간대에 다른스케줄이 존재하여 등록할 수 없습니다.");
+                    }
+                }
+                // 업무 번호 검삭
+                if (occupationRepository.findByNoAndBranchNo(requestDto.getOccupation().getNo(), requestDto.getBranch().getNo()) != null){
+                    for (ScheduleColor color : ScheduleColor.values()) {
+                        if (color.getColorCode().equals(requestDto.getColor())) {
+                            return scheduleRepository.save(requestDto.toEntity()).getNo();
+                        }
+                    }
+
+                    new IllegalArgumentException("등록할 수 없는 색상입니다.");
+                }else {
+                    new IllegalArgumentException("업무 명이 잘못되었습니다.");
+                }
+            }
+            new IllegalArgumentException("등록할 수 없는 스케줄입니다.");
+        }
+        new IllegalArgumentException("종료시간이 시작시간보다 빠른 시각입니다.");
+        return (long)0;
     }
 
     @Transactional(readOnly = true)
     @Override
     public List<ScheduleListResponseDto> getScheduleList(long no) {
-    return scheduleRepository.findAllBranchNo(no).stream()
-        .map(ScheduleListResponseDto::new)
-        .collect(Collectors.toList());
+        return scheduleRepository.findAllBranchNo(no).stream()
+                .map(ScheduleListResponseDto::new)
+                .collect(Collectors.toList());
     }
 
     @Transactional(readOnly = true)
@@ -48,8 +98,21 @@ public class ScheduleServiceImpl implements  ScheduleService {
         Schedule entity = scheduleRepository.findById(no)
                 .orElseThrow(() -> new IllegalArgumentException("해당 스케줄이 없습니다."));
 
-        entity.update(requestDto.getMember(), requestDto.getWorkStartTime(),
-                requestDto.getWorkEndTime(), requestDto.getOccupation(), requestDto.getColor());
+        if (requestDto.getWorkStartTime().isBefore(requestDto.getWorkEndTime())) {
+            if (occupationRepository.findByNoAndBranchNo(requestDto.getOccupation().getNo(), requestDto.getOccupation().getBranch().getNo()) != null) {
+                for (ScheduleColor color : ScheduleColor.values()) {
+                    if (color.getColorCode().equals(requestDto.getColor())) {
+                        for (ScheduleState state : ScheduleState.values()) {
+                            if (state.getState().equals(requestDto.getState())) {
+                                entity.update(requestDto.getMember(), requestDto.getWorkStartTime(),
+                                        requestDto.getWorkEndTime(), requestDto.getOccupation(), requestDto.getColor());
+                                return no;
+                            }
+                        } new IllegalArgumentException("잘못된 스케줄 상태입니다.");
+                    }
+                } new IllegalArgumentException("잘못된 색상입니다.");
+            } new IllegalArgumentException("잘못된 업무입니다.");
+        } new IllegalArgumentException("종료시간이 시작시간보다 빠른 시각입니다.");
 
         return no;
     }
@@ -59,7 +122,12 @@ public class ScheduleServiceImpl implements  ScheduleService {
     public void removeSchedule(long no) {
         Schedule entity = scheduleRepository.findById(no)
                 .orElseThrow(() -> new IllegalArgumentException("해당 스케줄이 없습니다."));
-
-        scheduleRepository.delete(entity);
+        if (!"W".equals(entity.getState())
+                && !"L".equals(entity.getState())
+                && !"A".equals(entity.getState())) {
+            scheduleRepository.delete(entity);
+        } else {
+            new IllegalArgumentException("출근, 지각, 결근 상태인 스케줄은 삭제할 수 없습니다.");
+        }
     }
 }
